It seems there might be a misunderstanding. The changes in the diff do not represent a replacement of "zap" with "slog". Instead, the diff indicates the Go version has been updated from 1.20 to 1.21 in GitHub workflow files. Here is a suggested correction:
Commit Message:
Update Go version to 1.21 in GitHub workflows

The Go version used in our GitHub workflows, including golangci-lint.yml and test.yml, has been updated from 1.20 to 1.21. This is to ensure that our tools and tests are running on the latest stable version and to take advantage of any performance benefits or new features available in version 1.21.