Update to Go v1.21, replace zap with slog for logging (#108)
* Update to Go v1.21, replace zap with slog for logging

* It seems there might be a misunderstanding. The changes in the diff do not represent a replacement of "zap" with "slog". Instead, the diff indicates the Go version has been updated from 1.20 to 1.21 in GitHub workflow files. Here is a suggested correction:

Commit Message:
Update Go version to 1.21 in GitHub workflows

The Go version used in our GitHub workflows, including golangci-lint.yml and test.yml, has been updated from 1.20 to 1.21. This is to ensure that our tools and tests are running on the latest stable version and to take advantage of any performance benefits or new features available in version 1.21.

* Update logging packages and upgrade Go version

This commit removes dependency on `go.uber.org/zap` and replaces its usage in logging middleware with `slog`. Also, added a new dependency `github.com/kr/text` for text processing.
In addition, updated Go and `github.com/GaijinEntertainment/go-exhaustruct` versions in `tools/go.mod`, which are necessary for tooling upgrades that help us maintain updated dependencies or latest Go features. Finally, various indirect dependencies were updated to their newer versions as per `go mod tidy`, ensuring we use the latest and secure versions.

* Update golangci-lint version in GitHub workflow

This commit updates the version of golangci-lint being used in the GitHub workflow from v1.53 to v1.54.1. This is done to take advantage of the improvements and bug fixes made in the latest version of golangci-lint.