Update dependencies and app version
Updated the versions of multiple dependencies in go.mod and go.sum after a round of package updates. This was done to keep the project up-to-date with the latest bug fixes and features provided by these dependencies. A new app version v0.5.0 has also been set in the Chart.yaml file to reflect these changes.