Add goleak package and implement graceful shutdown test
In this commit, we've added the `goleak` package to our `go.mod` file. This package is used to ensure no goroutines are leaked during the shutdown of our application.

We've also added a new test case in `main_test.go`. This test case, `TestGracefullShutdown`, checks if our application can be successfully shut down when it receives a `SIGTERM` signal.

We simulated this signal in our test by calling `syscall.Kill` after the application is up and running. By using this approach, we can verify that our application handles these signals as expected and cleans up necessary resources before shutting down.